### 2021-02-05 #3011 @whedon
[PRE REVIEW]: kallisto: A command-line interface to simplify 2 computational modelling and the generation of atomic features.
**Submitting author:** @f3rmion (<a href="http://orcid.org/0000-0002-3985-595X">Eike Caldeweyher</a>)
**Repository:** <a href="https://github.com/AstraZeneca/kallisto" target ="_blank">https://github.com/AstraZeneca/kallisto</a>
**Version:** v1.0.1
