### 2021-02-05 3957969
Remove backtics from references in paper.md. Pull citations into the preceding sentence.
