### 2021-02-05 #3011 @f3rmion
> preceding sentence

Hi @kyleniemeyer, I fixed the references.
According to the Whedon preview, the references seems to be correct now.
