### 2021-02-06 #3011 @richardjgowers
Hi @f3rmion I'm in the process of getting reviewers for this. One thing I noticed is you've missed that rdkit has had as extended Heuckel method available since 2019 (https://rdkit.blogspot.com/2019/06/doing-extended-hueckel-calculations.html)  Obviously not the best level of theory but worth a mention.
