### 2021-02-06 26930a8
Add EHT reference for RDKit to paper.md.
