+++ paper.bib
@@ -40,6 +40,14 @@
   year={2006}
 }
 
+@online{landrum2019,
+  author = {Landrum, Greg},
+  title = {YAeHMOP: Yet Another extended Hueckel Molecular Orbital Package},
+  year = {2019},
+  url = {https://github.com/greglandrum/yaehmop},
+  urldate = {2021-02-06}
+}
+
 @incollection{parr1980,
   title={Density functional theory of atoms and molecules},
   author={Parr, Robert G},
