### 2021-02-06 d46e9e4
Fix Hueckel in paper.md
