### 2021-02-08 0a953d8
Add link to GitBook repository in README.md
