### 2021-02-11 4a2df26
Refactor kallisto package. Release at PyPI v1.0.2. Closes #2
