### 2021-02-12 d2368c5
Increase unit test coverage for atom. Closes #3
