### 2021-02-16 3aa9f97
Increase unit test coverage for molecule. Close #4
