### 2021-02-18 ff5c2a3
Increase unit test coverage for strucreader. Close #5
