+++ tests/test_strucreader.py
@@ -3,6 +3,8 @@
 import os
 import tempfile
 
+import pytest
+
 from kallisto.atom import Atom
 import kallisto.reader.strucreader as ksr
 
@@ -11,7 +13,7 @@ s = os.linesep
 
 
 # turbomole coord files can be read
-def test_a_user_can_read_a_TMfile():
+def test_a_user_can_read_coord():
     with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
         f.write("$coord" + s)
         f.write("  1.87167924 -0.101043656  0.1596818582  c" + s)
@@ -26,3 +28,16 @@ def test_a_user_can_read_a_TMfile():
         got = type(atoms[0])
         want = Atom
         assert got is want
+
+
+def test_a_user_cannot_read_invalid_coord():
+    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
+        f.write("$xoord" + s)
+        f.write("  1.87167924 -0.101043656  0.1596818582  c" + s)
+        f.write("$end")
+        f.flush()
+        fname = open(f.name, "r+")
+        with pytest.raises(SystemExit) as error:
+            ksr.read(fname)
+        assert error.value.args[0] == 1
+        fname.close()
