+++ src/kallisto/console.py
@@ -57,7 +57,7 @@ def cli(config, verbose: bool, shift: int):
 def cns(config, inp: str, out: click.File, cntype: str):
     """Atomic coordination numbers."""
 
-    molecule = ksr.constructMolecule(geometry=inp)
+    molecule = ksr.constructMolecule(geometry=inp, out=out)
     cns = molecule.get_cns(cntype)
     verbosePrinter(config.verbose, cns, out)
 
@@ -85,7 +85,7 @@ def cns(config, inp: str, out: click.File, cntype: str):
 def cnsp(config, inp: str, out: click.File, cntype: str):
     """Atomic coordination number spheres."""
 
-    molecule = ksr.constructMolecule(geometry=inp)
+    molecule = ksr.constructMolecule(geometry=inp, out=out)
     nat = molecule.get_number_of_atoms()
     cnsp = molecule.get_cnspheres(cntype)
     for i in range(nat):
@@ -119,7 +119,7 @@ def bonds(config, inp: str, partner: int, constrain: bool, out: click.File):
 
     import os
 
-    molecule = ksr.constructMolecule(geometry=inp)
+    molecule = ksr.constructMolecule(geometry=inp, out=out)
 
     if partner == "X":
         # Get index table of covalent bonding partners
@@ -174,14 +174,14 @@ def sort(config, inp: str, start: str, out: click.File):
     start defines on which atom we start the sorting process.
     """
 
-    molecule = ksr.constructMolecule(geometry=inp)
+    molecule = ksr.constructMolecule(geometry=inp, out=out)
     bonds = molecule.get_bonds()
     nat = molecule.get_number_of_atoms()
 
     # construct graph
     from kallisto.sort import Graph
 
-    g = Graph(inp)
+    g = Graph(inp, out)
 
     for i in range(nat):
         partners = bonds[i]
@@ -218,7 +218,7 @@ def sort(config, inp: str, start: str, out: click.File):
 def eeq(config, inp: str, out: click.File, chrg: int):
     """Electronegativity equilibration atomic partial charges."""
 
-    molecule = ksr.constructMolecule(geometry=inp)
+    molecule = ksr.constructMolecule(geometry=inp, out=out)
     nat = molecule.get_number_of_atoms()
     eeq = molecule.get_eeq(chrg)
     for i in range(nat):
@@ -248,7 +248,7 @@ def eeq(config, inp: str, out: click.File, chrg: int):
 def alp(config, inp: str, out: click.File, chrg: int, molecular: bool):
     """Static atomic polarizabilities in Bohr^3."""
 
-    molecule = ksr.constructMolecule(geometry=inp)
+    molecule = ksr.constructMolecule(geometry=inp, out=out)
     nat = molecule.get_number_of_atoms()
     alp = molecule.get_alp(charge=chrg)
     if molecular:
@@ -291,7 +291,7 @@ def alp(config, inp: str, out: click.File, chrg: int, molecular: bool):
 def vdw(config, inp: str, out: click.File, chrg: int, vdwtype: str, angstrom: bool):
     """Charge-dependent atomic van der Waals radii in Bohr."""
 
-    molecule = ksr.constructMolecule(geometry=inp)
+    molecule = ksr.constructMolecule(geometry=inp, out=out)
     nat = molecule.get_number_of_atoms()
     vdw = np.zeros(shape=(nat,), dtype=np.float64)
 
@@ -327,20 +327,19 @@ def rms(config, compare: Tuple[str, str], out: click.File):
 
     from kallisto.rmsd import rmsd
 
-    mol1 = ksr.constructMolecule(geometry=compare[0])
[…]
