### 2021-02-21 08176ed
Add DOIs to references in paper.bib
