+++ paper.bib
@@ -2,6 +2,7 @@
   title={Understanding the basics of QSAR for applications in pharmaceutical sciences and risk assessment},
   author={Roy, Kunal and Kar, Supratik and Das, Rudra Narayan},
   year={2015},
+  DOI={10.1016/c2014-0-00286-9},
 }
 
 @article{sprous2010,
@@ -12,6 +13,7 @@
   number={6},
   pages={619--637},
   year={2010},
+  DOI={10.2174/156802610791111506},
 }
 
 @article{dudek2006,
@@ -22,6 +24,7 @@
   number={3},
   pages={213--228},
   year={2006},
+  DOI={10.2174/138620706776055539},
 }
 
 @article{verma2010,
@@ -32,6 +35,7 @@
   number={1},
   pages={95--115},
   year={2010},
+  DOI={10.2174/156802610790232260},
 }
 
 @article{landrum2006,
@@ -64,6 +68,7 @@
   number={5},
   pages={1253},
   year={1999},
+  DOI={10.1103/revmodphys.71.1253},
 }
 
 @article{dreuw2005,
@@ -74,6 +79,7 @@
   number={11},
   pages={4009--4037},
   year={2005},
+  DOI={10.1021/cr0505627},
 }
 
 @article{adamo1999,
@@ -84,6 +90,7 @@
   number={13},
   pages={6158--6170},
   year={1999},
+  DOI={10.1063/1.478522},
 }
 
 @article{weigend2005,
@@ -94,6 +101,7 @@
   number={18},
   pages={3297--3305},
   year={2005},
+  DOI={10.1039/b508541a},
 }
 
 @article{caldeweyher2019,
@@ -104,6 +112,7 @@
   number={15},
   pages={154122},
   year={2019},
+  DOI={10.1063/1.5090222},
 }
 
 @article{hirshfeld1977,
@@ -114,6 +123,7 @@
   number={2},
   pages={129--138},
   year={1977},
+  DOI={10.1007/bf00549096},
 }
 
 @article{bannwarth2020,
@@ -122,4 +132,5 @@
   journal={WIREs Comput. Mol. Sci.},
   pages={e01493},
   year={2020},
+  DOI={10.1002/wcms.1493},
 }
