### 2021-02-21 eba10f9
Add JOSS badge to README.md
