### 2021-02-23 #3050 @f3rmion
> Hi @f3rmion . In the `README` (under "Full Author List") your name is listed alongside "Philipp Pracht" but you are the only author in the JOSS article. I can see Pracht's contributions are only acknowledged in the paper; could you please confirm that this is intended, so that I can tick off the "Contribution and authorship" box?
>
> I'm asking this since there are only 7 commits in the public repository, therefore it is impossible to see the different contributions.
Confirmed, I talked to him and he only wants to be mentioned within the acknowledgements.
