### 2021-02-27 AstraZeneca/kallisto#8 @RMeli
[JOSS] Installation instructions
Installation with `pip` works flawlessly in a clean `conda` environment.

Installation from source also works flawlessly, but I think the instructions could be simplified so that users can choose how to manage their environments. For example, `poetry` is also available on the `conda-forge` channel and `nox` can be installed using `conda` instead of `pip`.

I suggest to decouple the installation instructions for of Python and dependencies from the installation of `kallisto` as follows:

- [x] Clearly state the dependencies to install from source (`poetry`, Python version, ...)
    - [x] Link to the installation instructions of such dependencies
- [x] Simplify the instructions to install `kallisto`: clone, change directory, call potery
- [x] Create a separate section for testing
    - [x] List dependencies for testing (`nox`)

The advantage of linking to the installation instructions of the dependencies, IMHO, is that you won't need to maintain such instructions up to date yourself and users can install such dependencies in their preferred way (`conda` vs `pip`, ...).

Additional comments/questions:

- [x] Is "The Verbose Mode" section intended to be in the installation page?
- [ ] It would also be nice to have a list of the main dependencies at the top of the installation page.

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
