### 2021-02-28 AstraZeneca/kallisto#8 @f3rmion
Hi @RMeli,

This is a very good suggestion. I have already changed the [kallisto docs](https://ehjc.gitbook.io/kallisto/setup) accordingly.
Now all dependencies are listed and the installation process from the source is simplified. I also added a [testing suite](https://ehjc.gitbook.io/kallisto/test) section that lists all nox sessions for unit test purposes. I furthermore shifted the "Verbose mode" and "Need help" to [misc](https://ehjc.gitbook.io/kallisto/misc/help). Once you approve this, I will also update the README.md within the kallisto repository.
