### 2021-02-28 806c2e1
Fix most alerts from lgtm.com. Increase version number to v1.0.4
