+++ tests/test_version.py
@@ -4,4 +4,4 @@ from kallisto import __version__
 
 
 def test_version():
-    assert __version__ == "1.0.3"
+    assert __version__ == "1.0.4"
