### 2021-03-01 7d02fdf
Fix two more lgtm.com alerts.
