### 2021-03-01 AstraZeneca/kallisto#8 @RMeli
@f3rmion Looks very good now, I like the separated and expanded section for the test suite! The last comment I have is about the list of dependencies: I think it is a bit difficult to read now and it seems that most dependencies are related to development.

- [ ] Would it be possible to just list the main dependencies needed to run the scripts?

You could link to the `project.toml` file for further details, where the distinction between dependencies and development dependencies is very clear (on this note: is `flake8` supposed to be under both `[tool.poetry.dependencies]` and `[tool.poetry.dev-dependencies]`?).
