### 2021-03-02 d8420d8
Add simplified installation instructions to README.md.
