+++ README.md
@@ -5,31 +5,40 @@
 ##
 
 [![Documentation](https://img.shields.io/badge/GitBook-Docu-lightgrey)](https://app.gitbook.com/@ehjc/s/kallisto/)
-[![License](https://img.shields.io/badge/license-Apache%202-blue)](https://img.shields.io/badge/license-Apache%202-blue)
 [![Maturity Level](https://img.shields.io/badge/Maturity%20Level-ML--1-orange)](https://img.shields.io/badge/Maturity%20Level-ML--1-orange)
 [![Tests](https://github.com/AstraZeneca/kallisto/workflows/Tests/badge.svg)](https://github.com/AstraZeneca/kallisto/actions?workflow=Tests)
 [![codecov](https://codecov.io/gh/AstraZeneca/kallisto/branch/master/graph/badge.svg?token=HI0U0R96X8)](https://codecov.io/gh/AstraZeneca/kallisto)
 [![Language grade: Python](https://img.shields.io/lgtm/grade/python/g/AstraZeneca/kallisto.svg?logo=lgtm&logoWidth=18)](https://lgtm.com/projects/g/AstraZeneca/kallisto/context:python)
 [![status](https://joss.theoj.org/papers/16126cbcfb826bf4810d243a009a6b02/status.svg)](https://joss.theoj.org/papers/16126cbcfb826bf4810d243a009a6b02)
 
-Table of Contents
------------------
+# Table of Contents
 
 - Full Author List
 - Introduction
 - Installation
+- Testing suite
 - Reference
 
-Full Author List
-----------------
+# Full Author List
 
 Eike Caldeweyher and Philipp Pracht
 
-Introduction
-------------
+# Introduction
 
 We developed the `kallisto` program for the efficient and robust calculation of atomic features using molecular geometries either in a ``xmol`` or a ``Turbomole`` format.
 Furthermore, several modelling tools are implemented, e.g., to calculate root-mean squared deviations via quaternions (including rotation matrices), sorting of molecular geometries and many more. All features of ``kallisto`` are described in detail within our [documentation](https://app.gitbook.com/@ehjc/s/kallisto/) ([GitBook repository](https://github.com/f3rmion/gitbook-kallisto)).
+
+Main dependencies
+-----------------
+
+```bash
+click 7.1.2 Composable command line interface toolkit
+numpy 1.20.1 NumPy is the fundamental package for array computing with Python.
+scipy 1.6.0 SciPy: Scientific Library for Python
+└── numpy >=1.16.5
+```
+
+For a list of all dependencies have a look at the pyproject.toml file.
 
 Installation from PyPI
 ----------------------
@@ -42,54 +51,103 @@ pip install kallisto
 Installation from Source
 ------------------------
 
-`kallisto` runs on `python3`
+Requirements to install ``kallisto``from sources:
+- [poetry](https://python-poetry.org/docs/#installation)
+- [pyenv](https://github.com/pyenv/pyenv#installation) or [conda](https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html)
+- python >=3.7
 
-Python development setup. Install the `pyenv` python version manager:
+First check that ``poetry`` is running correctly (v1.0.10 at the time of writing)
+
 ```bash
-curl https://pyenv.run | bash
-```
-and add this to the `~/.bashrc` and source it:
-```bash
-export PATH="~/.pyenv/bin:$PATH"
-eval "$(pyenv init -)"
-eval "$(pyenv virtualenv-init -)"
-```
-Install the latest python versions:
-```bash
-pyenv install 3.8.2
-pyenv install 3.7.7
-pyenv local 3.8.2 3.7.7
+> poetry --version
+Poetry version 1.0.10
 ```
 
-Now we are ready to set up `kallisto`.
[…]
