+++ pyproject.toml
@@ -14,7 +14,6 @@ python = "^3.7"
 click = "^7.1.2"
 numpy = "^1.19.0"
 scipy = "^1.5.2"
-flake8 = "^3.8.4"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
