### 2021-03-02 AstraZeneca/kallisto#8 @f3rmion
@RMeli I further condensed the dependencies to the core (main dependencies) and removed ``flake8`` from ``tool.poetry.dependencies`` (d8420d8).
