### 2021-03-02 #3050 @f3rmion
@RMeli  Installation instructions have been simplified both in the [docs](https://app.gitbook.com/@ehjc/s/kallisto/setup) and in the [README.md](https://github.com/AstraZeneca/kallisto/blob/master/README.md). Thanks!
