### 2021-03-03 948e9a8
Create CODE_OF_CONDUCT.md
