### 2021-03-06 #3050 @Sulstice
Ah I am catching up on the threads. @f3rmion  I was reading the README first and then comments. I've begun the review and looking through the docs first. I can close some of the issues if you want me to.

For your gitbook - is there a way.I can make suggestion edits in there? or should I continue with github issues?

Apologies for taking some time to respond - weird situation over here in texas...haha
