+++ src/grid.py
@@ -4,13 +4,18 @@ from collections import Counter
 
 import numpy as np
 
-# Lebedev-Laikov angular grids
+# Angular Lebedev-Laikov grids (6-48 equivalent points)
+#
+# Original (C): Dmitri N. Laikov
+# Adapted (Fortran): Christoph van Wuellen (CvW)
+# Translated into Python by permission of CvW
 #
 # Reference:
 # V.I. Lebedev, and D.N. Laikov, "A quadrature formula for the sphere of the
 # 131st algebraic order of accuracy", Doklady Mathematics, Vol. 59, No. 3,
 # 1999, pp. 477-481.
 
+# Define grid sizes
 gridSize = dict(
     [
         (0, 6),
@@ -27,50 +32,50 @@ gridSize = dict(
     ]
 )
 
-# zero as constant to keep Oh generators clean
+# keep octahedral symmetry
 z = 0.0
 
 
-def getAngularLebedevGrid(nAng: int):
-    """Wrapper for generation of angular Lebedev-Laikov grids.
+def getLebedevLaikovGrid(n: int):
+    """Generate angular Lebedev-Laikov grid.
 
     Parameter:
-     nAng - size of the grid
-     grid - contains grid points on output
-     weights - contains weights on output"""
+     n - size of grid
+     grid - grid points
+     weights - point weights"""
 
     # initialize count
     count = Counter()  # type: ignore
     count["points"] = 0
 
-    size = gridSize[nAng]
+    size = gridSize[n]
     grid = np.zeros(shape=(size, 3), dtype=np.float64)
     weights = np.zeros(shape=(size,), dtype=np.float64)
 
-    if nAng == 0:
+    if n == 0:
         grid, weights, count = ld0006(grid, weights, count)
-    elif nAng == 1:
+    elif n == 1:
         grid, weights, count = ld0014(grid, weights, count)
-    elif nAng == 2:
+    elif n == 2:
         grid, weights, count = ld0026(grid, weights, count)
-    elif nAng == 3:
+    elif n == 3:
         grid, weights, count = ld0038(grid, weights, count)
-    elif nAng == 4:
+    elif n == 4:
         grid, weights, count = ld0050(grid, weights, count)
-    elif nAng == 5:
+    elif n == 5:
         grid, weights, count = ld0074(grid, weights, count)
-    elif nAng == 6:
+    elif n == 6:
         grid, weights, count = ld0086(grid, weights, count)
-    elif nAng == 7:
+    elif n == 7:
         grid, weights, count = ld0110(grid, weights, count)
-    elif nAng == 8:
+    elif n == 8:
         grid, weights, count = ld0146(grid, weights, count)
-    elif nAng == 9:
+    elif n == 9:
         grid, weights, count = ld0170(grid, weights, count)
     else:
         grid, weights, count = ld0194(grid, weights, count)
 
-    if count["points"] != gridSize[nAng]:
+    if count["points"] != gridSize[n]:
         raise RuntimeError(
[…]
