### 2021-03-06 bec5949
Add python version badge to README.md. Close #11
