### 2021-03-06 34f5da2
Link developers in README.md. Close #12
