### 2021-03-06 #3050 @f3rmion
@Sulstice I closed all issues:
- added a python 3.7 badge
- linked all developers in README.md
- added @pprcht to the pyproject.toml
