### 2021-03-06 AstraZeneca/kallisto#15 @RMeli
[JOSS] Paper
### Content

While the "Statement of Need" in the manuscript introduces the need of computation of atomic features, it is not clear exactly what the actual capabilities of the software are. I understand the main functionality is described in the online documentation, but I think it should also be clearer from the paper. This will allow to add proper references to the implemented methods as well (if relevant).

The same applies to the mentioned "computational modelling helpers". It would be nice to know which one they are in the paper (at least the most useful/relevant ones).

I also think that specifying DFT functionals and basis sets is not relevant in the paper and would be better off left to the documentation (where I can't see any mention of this?). I think `(e.g., static/dynamic poralizabilities with time-dependent DFT data)` would be much more readable than `(e.g., static/dynamic polarizabilitities with time-dependent
(Dreuw & Head-Gordon, 2005) PBE0 (Adamo & Barone, 1999) / d-aug-def2-QZVP (Weigend
& Ahlrichs, 2005) data)`.

### Typos

- Line 9: `usecases` -> `use cases`
- Line 30: `polarizabilitities` -> `polarizabilities`
- Line 38: `Stefan Grimme and co-wo rker` -> `Stefan Grimme and co-workers`

(I suck at spotting typos, so I might add more later on...)

### References

I think the phrase

> Tasks as, e.g., building predictive models, performing virtual screening, or predicting compound activities are potential use cases for such ML applications.

could do with some references about ML models for VS and activity prediction.

Multiple references are not correctly displayed. As it can be inferred  from the [example paper and bibliography](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) (not very clear , I know...!), multiple references can be put together with the following syntax: `[@ref1; @ref2]`.

### Suggestions

- [x] Add details of implemented featurizers; this can be a comma-separated list
    - [x] Add relevant references (if applicable)
- [x] Add list of useful computational modelling helpers
    - [x] Add relevant references (if applicable)
- [x] Fix typos
- [x] Combine multiple reference into a single one, where necessary
- [x] Move low-level details to documentation (DFT functionals and basis sets)

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
