### 2021-03-07 #3050 @Sulstice
@f3rmion Awesome, I've got more incoming as I go line by line through your stuff.
