### 2021-03-08 2471cb4
Remove verbose mode to make CLI more intuitive (#14)
