+++ tests/test_cli.py
@@ -32,14 +32,16 @@ def test_cli_print_help(runner):
 
 
 # test cli part for coordination number
-def test_cli_cns_exp_no_verbose(runner):
+def test_cli_cns_exp_silent(runner):
     with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
         f.write("$coord" + s)
         f.write("0.00000000000000 0.00000000000000 -1.06176434496059 c" + s)
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["cns", "--inp", f.name, "--cntype", "exp"])
+        result = runner.invoke(
+            cli, ["--silent", "cns", "--inp", f.name, "--cntype", "exp"]
+        )
         assert result.exit_code == 0
 
 
@@ -50,9 +52,7 @@ def test_cli_cns_exp(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(
-            cli, ["--verbose", "cns", "--inp", f.name, "--cntype", "exp"]
-        )
+        result = runner.invoke(cli, ["cns", "--inp", f.name, "--cntype", "exp"])
         assert result.exit_code == 0
         assert "0.98678927" in result.output
 
@@ -64,11 +64,9 @@ def test_cli_cns_cov(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(
-            cli, ["--verbose", "cns", "--inp", f.name, "--cntype", "cov"]
-        )
+        result = runner.invoke(cli, ["cns", "--inp", f.name, "--cntype", "cov"])
         assert result.exit_code == 0
-        assert "0.91894762" in result.output
+        assert "0.9189476" in result.output
 
 
 def test_cli_cns_erf(runner):
@@ -78,22 +76,20 @@ def test_cli_cns_erf(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(
-            cli, ["--verbose", "cns", "--inp", f.name, "--cntype", "erf"]
-        )
+        result = runner.invoke(cli, ["cns", "--inp", f.name, "--cntype", "erf"])
         assert result.exit_code == 0
-        assert "0.98784659" in result.output
+        assert "0.9878465" in result.output
 
 
 # test cli part for bonds
-def test_cli_bonds_no_verbose(runner):
+def test_cli_bonds_silent(runner):
     with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
         f.write("$coord" + s)
         f.write("0.00000000000000 0.00000000000000 -1.06176434496059 c" + s)
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["bonds", "--inp", f.name])
+        result = runner.invoke(cli, ["--silent", "bonds", "--inp", f.name])
         assert result.exit_code == 0
 
 
@@ -104,7 +100,7 @@ def test_cli_bonds(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["--verbose", "bonds", "--inp", f.name])
+        result = runner.invoke(cli, ["bonds", "--inp", f.name])
         assert result.exit_code == 0
 
 
@@ -115,9 +111,7 @@ def test_cli_bonds_partner(runner):
[…]
