### 2021-03-08 7e8f24a
Explicitly list supported Python versions in README.md (#11)
