### 2021-03-08 f962e12
Resolve suggestions from JOSS review (#15)
