+++ paper.bib
@@ -1,136 +1,200 @@
+@article{grimme2010,
+  url={https://doi.org/10.1063/1.3382344},
+  Title={{A consistent and accurate ab initio parametrization of density functional dispersion correction (DFT-D) for the 94 elements H-Pu}},
+  Author={{Grimme}, S. and {Antony}, J. and {Ehrlich}, S. and {Krieg}, H.},
+  Journal={J. Chem. Phys.},
+  Volume={132},
+  Number={15},
+  Pages={154104},
+  Year={2010},
+  DOI={10.1063/1.3382344}
+}
+
+@article{caldeweyher2019,
+  url={https://doi.org/10.1063/1.5090222},
+  Title={{A generally applicable atomic-charge dependent London dispersion correction}},
+  Author={{Caldeweyher}, E. and {Ehlert}, S. and {Hansen}, A. and {Neugebauer}, H. and {Spicher}, S. and {Bannwarth}, C. and {Grimme}, S.},
+  Journal={J. Chem. Phys.},
+  Volume={150},
+  Number={15},
+  Pages={154122},
+  Year={2019},
+  DOI={10.1063/1.5090222},
+}
+
+@article{fedorov2018,
+  url={https://doi.org/10.1103/PhysRevLett.121.183401},
+  Title={{Quantum-mechanical relation between atomic dipole polarizability and the van der Waals radius}},
+  Author={{Fedorov}, D. V. and {Sadhukhan}, M. and {St{\"o}hr}, M. and {Tkatchenko}, A.},
+  Journal={Phys, Rev. Lett.},
+  Volume={121},
+  Number={18},
+  Pages={183401},
+  Year={2018},
+  DOI={10.1103/PhysRevLett.121.183401},
+}
+
+@article{rahm2017,
+  url={https://doi.org/10.1002/chem.201700610},
+  Title={{Corrigendum: Atomic and Ionic Radii of Elements 1--96}},
+  Author={{Rahm}, M. and {Hoffmann}, R. and {Ashcroft}, N. W.},
+  Journal={Chem. Eur. J.},
+  Volume={23},
+  Number={16},
+  Pages={4017--4017},
+  Year={2017},
+  DOI={10.1002/chem.201700610},
+}
+
+@article{mantina2009,
+  url={https://doi.org/10.1021/jp8111556},
+  Title={{Consistent van der Waals radii for the whole main group}},
+  Author={{Mantina}, M. and {Chamberlin}, A. C. and {Valero}, R. and {Cramer}, C. J. and {Truhlar}, D. G.},
+  Journal={J. Phys. Chem. A},
+  Volume={113},
+  Number={19},
+  Pages={5806--5812},
+  Year={2009},
+  DOI={10.1021/jp8111556},
+}
+
+@article{brethome2019,
+  url={https://doi.org/10.1021/acscatal.8b04043},
+  Title={{Conformational Effects on Physical-Organic Descriptors: The Case of Sterimol Steric Parameters}},
+  Author={{Brethome}, A. V. and {Fletcher}, S. P. and {Paton}, R. S.},
+  Journal={ACS Catalysis},
+  Volume={9},
+  Number={3},
+  Pages={2313--2323},
+  Year={2019},
+  DOI={10.1021/acscatal.8b04043},
+}
+
 @book{roy2015,
-  title={Understanding the basics of QSAR for applications in pharmaceutical sciences and risk assessment},
-  author={Roy, Kunal and Kar, Supratik and Das, Rudra Narayan},
-  year={2015},
+  url={https://doi.org/10.1016/c2014-0-00286-9},
+  Author={{Roy}, K. and {Kar}, S. and {Das}, R. N.},
[…]
