### 2021-03-08 AstraZeneca/kallisto#15 @f3rmion
**Add details of implemented featurizers; this can be a comma-separated list**
I added a small list of all implemented atomic features with relevant references.
- Coordination number: [1](https://doi.org/10.1063/1.3382344), [2](https://doi.org/10.1063/1.5090222)
- Electronegativity equilibration atomic partial charges: [1](https://doi.org/10.1063/1.5090222)
- Dynamic polarzabilities: [1](https://doi.org/10.1063/1.3382344), [2](https://doi.org/10.1063/1.5090222)
- Van-der-Waals radii: [1](https://doi.org/10.1103/PhysRevLett.121.183401), [2](https://doi.org/10.1002/chem.201700610), [3](https://doi.org/10.1021/jp8111556)

**Add list of useful computational modelling helpers**
I also added a list of all implemented modelling helpers (without references).

**Fix typos**
Done.

**Combine multiple references into a single one, where necessary**
Done. I tried to stick to the format of the references in your resource.

**Move low-level details to documentation (DFT functionals and basis sets)**
I removed the actual computational details from the paper and added them to the [docs](https://app.gitbook.com/@ehjc/s/kallisto/features/alp).
