### 2021-03-08 #3050 @f3rmion
@Sulstice OK, I'll check your issues within the repo :)

@RMeli First commit to resolve all suggestions in ``paper.md`` and ``paper.bib``
