### 2021-03-08 AstraZeneca/kallisto#15 @RMeli
Thanks @f3rmion . I think the two lists are really useful in the paper; can you pleas move the after the statement of need?

Just to confirm, there are no references for the implemented modelling helpers? I'm asking because, for example, fast methods for RMSD calculations are usually published.
