+++ paper.bib
@@ -70,6 +70,18 @@
   DOI={10.1021/acscatal.8b04043},
 }
 
+@article{coutsias2004,
+  url={https://doi.org/10.1002/jcc.20110},
+  Title={{Using quaternions to calculate RMSD}},
+  Author={{Coutsias}, E. A. and {Seok}, C. and {Dill}, K. A.},
+  Journal={J. Comput. Chem.},
+  Volume={25},
+  Number={15},
+  Pages={1849--1857},
+  Year={2004},
+  DOI={10.1002/jcc.20110}
+}
+
 @book{roy2015,
   url={https://doi.org/10.1016/c2014-0-00286-9},
   Author={{Roy}, K. and {Kar}, S. and {Das}, R. N.},
