### 2021-03-09 AstraZeneca/kallisto#15 @f3rmion
Hi @RMeli,

thanks for the comments! I shifted both lists and added an RMSD reference.
For the other modelling helpers, I don't really have references to cite, but I know at least for the structure exchange that there are similar approaches available, however, I guess they are based on a different approach using template structures.
