### 2021-03-09 15a097b
Prepare release and bump up version number to v1.0.5
