### 2021-03-11 9d47012
Add permission note to sterics.
