### 2021-03-12 AstraZeneca/kallisto#19 @RMeli
[JOSS] grid module outside of kallisto
I started looking at the source code and I noticed that the `grid.py` module is outside the `kallisto` package. When `kallisto` is imported, the `grid` module also becomes available, possibly clashing with or shadowing other modules. Is this intended?

I think it would be better to have the `grid` module within the package, unless I'm missing a good reason not to.

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
