### 2021-03-13 AstraZeneca/kallisto#19 @f3rmion
Hi @RMeli,

The Lebedev-Laikov grid is not yet available as an implemented feature. I would like to change this in the following release and also want to extend the grid further. Then I will also include it accordingly in the ``kallisto`` package as noted by you :)
