### 2021-03-14 AstraZeneca/kallisto#19 @RMeli
Thanks for the explanation @f3rmion. However, if I install `kallisto` from source, the `grid` module seem to be available:
```
>>> import grid
>>> dir(grid)
['Counter', '__builtins__', '__cached__', '__doc__', '__file__', '__loader__', '__name__', '__package__', '__spec__', 'genOh1', 'genOh2', 'genOh3', 'genOh4', 'genOh5', 'genOh6', 'getAngularLebedevGrid', 'gridSize', 'ld0006', 'ld0014', 'ld0026', 'ld0038', 'ld0050', 'ld0074', 'ld0086', 'ld0110', 'ld0146', 'ld0170', 'ld0194', 'np', 'z']
```

If grids are not yet available/fully implemented/supported, I think they should be removed from the installation, so that any confusion is avoided.
