### 2021-03-15 994d754
Use tuple for slots in atom (#20)
