### 2021-03-15 AstraZeneca/kallisto#23 @RMeli
[JOSS] RMSD
In the `rmsd.py` module, it seems that there is a lot of code that could be removed in favor of `scipy.spatial.transform` and `scipy.linalg`. My view here is that code from another library is code you don't have to maintain, and a library such as `scipy` is usually very reliable, extensively tested and possibly faster.

These are the functions that could use `scipy` functionality; it would remove a lot of lines of code:
* `rotionMatrix()` could use `scipy.spatial.transform.Rotation`
   * Build a `scipy.spatial.transform.Rotation` using `from_quat()`
   * Return corresponding rotation matrix with `as_matrix()`
* `dstmev()` could be substituted with a `scipy.linalg.eigh()`?
   * This should allow the removal of `givens4`
   * If there is a good reason to implement this for scratch please let me know, I definitely missed it
* `getRodriguezRotation()` could use `scipy.spatial.transform.Rotation` internally
   * Build a `scipy.spatial.transform.Rotation` using `from_rotvec()`
   * Apply the rotation with `apply()`
* `getRotationMatrix()` could use `scipy.spatial.transform.Rotation` internally
   * Build a `scipy.spatial.transform.Rotation` using `from_rotvec()`
   * Return corresponding rotation matrix with `as_matrix()`
* `getDist` could be vectorised or substituted with `scipy.spatial.distance.euclidean` (or similar)

Up to you if you want to change these, but I think it would be helpful for future development and maintenance. The best code is no code ;)

Additionally:
* [x] The function`pythag()` is missing a docstring.
* [x] Is the definition of `unit` in `getRodriguezRotation()` correct?

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
