### 2021-03-15 AstraZeneca/kallisto#19 @f3rmion
Hi @RMeli,

> Thanks for the explanation @f3rmion. However, if I install kallisto from source, the grid module seem to be available:

Yes, you are right it was available. I shifted the ``grid`` code to the ``kallisto`` package to overcome the clashing/shadowing problems and added unit tests respectively.

> If grids are not yet available/fully implemented/supported, I think they should be removed from the installation, so that any confusion is avoided.

In one of the next releases, I will further extend it with finer grids and add new features based on it, however, the implemented grid can be used as-is already now.
