### 2021-03-15 8d23d89
Change units default to CODATA2018 via SciPy (Close #21)
