+++ mypy.ini
@@ -1,5 +1,5 @@
 # mypy.ini
 [mypy]
 
-[mypy-nox.*,pytest,numpy,scipy,click]
+[mypy-nox.*,pytest,numpy,scipy.*,click]
 ignore_missing_imports = True
