### 2021-03-15 AstraZeneca/kallisto#23 @f3rmion
Hi @RMeli,

Thanks for the effort, I definitely appreciate it! Try to simplify the ``rmsd``module according to your suggestions :)
