### 2021-03-15 ddb0320
Use dict to store theta_b values for vdw (#22)
