+++ src/kallisto/data/vdw.py
@@ -2,199 +2,198 @@
 
 # Scaling factors to match vdw radii from Table 12
 # in J. Phys. Chem. A, Vol. 113, No. 19, 2009
-truhlar = [
-    0.65,  # H
-    1.00,  # He
-    0.65,  # Li
-    0.67,  # Be
-    0.92,  # B
-    0.89,  # C
-    0.87,  # N
-    0.89,  # O
-    0.91,  # F
-    1.00,  # Ne
-    0.81,  # Na
-    0.69,  # Mg
-    0.76,  # Al
-    0.93,  # Si
-    0.84,  # P
-    0.87,  # S
-    0.89,  # Cl
-    1.00,  # Ar
-    0.88,  # K
-    0.83,  # Ca
+truhlar = {
+    "H": 0.65,
+    "He": 1.00,
+    "Li": 0.65,
+    "Be": 0.67,
+    "B": 0.92,
+    "C": 0.89,
+    "N": 0.87,
+    "O": 0.89,
+    "F": 0.91,
+    "Ne": 1.00,
+    "Na": 0.81,
+    "Mg": 0.69,
+    "Al": 0.76,
+    "Si": 0.93,
+    "P": 0.84,
+    "S": 0.87,
+    "Cl": 0.89,
+    "Ar": 1.00,
+    "K": 0.88,
+    "Ca": 0.83,
     # Sc-Zn
     # interpolated wrt Z
     # Start: Ca factor
     # End:   Ga factor
-    0.82,  # Sc
-    0.82,  # Ti
-    0.82,  # V
-    0.81,  # Cr
-    0.81,  # Mn
-    0.80,  # Fe
-    0.80,  # Co
-    0.79,  # Ni
-    0.79,  # Cu
-    0.79,  # Zn
+    "Sc": 0.82,
+    "Ti": 0.82,
+    "V": 0.82,
+    "Cr": 0.81,
+    "Mn": 0.81,
+    "Fe": 0.80,
+    "Co": 0.80,
+    "Ni": 0.79,
+    "Cu": 0.79,
+    "Zn": 0.79,
     #
-    0.78,  # Ga
-    0.92,  # Ge
-    0.84,  # As
-    0.89,  # Se
-    0.88,  # Br
-    1.00,  # Kr
-    0.96,  # Rb
-    0.86,  # Sr
+    "Ga": 0.78,
+    "Ge": 0.92,
+    "As": 0.84,
+    "Se": 0.89,
+    "Br": 0.88,
+    "Kr": 1.00,
+    "Rb": 0.96,
+    "Sr": 0.86,
     # Y-Cd
     # interpolated wrt Z
     # Start: Sr factor
     # End:   In factor
-    0.86,  # Y
-    0.85,  # Zr
-    0.84,  # Nb
-    0.83,  # Mo
-    0.83,  # Tc
-    0.82,  # Ru
-    0.81,  # Rh
-    0.80,  # Pd
-    0.79,  # Ag
-    0.79,  # Cd
+    "Y": 0.86,
+    "Zr": 0.85,
+    "Nb": 0.84,
+    "Mo": 0.83,
+    "Tc": 0.83,
+    "Ru": 0.82,
+    "Rh": 0.81,
+    "Pd": 0.80,
+    "Ag": 0.79,
+    "Cd": 0.79,
     #
-    0.78,  # In
-    0.90,  # Sn
-    0.89,  # Sb
-    0.91,  # Te
-    0.90,  # I
-    1.00,  # Xe
-    1.04,  # Cs
-    0.89,  # Ba
-    0.87,  # La
[…]
