### 2021-03-16 f39b577
Remove getDist, simplify getRodriguezRotation + fix unit, remove getRotationMatrix, and add docstring in pythag (#23)
