### 2021-03-16 AstraZeneca/kallisto#23 @f3rmion
Hi @RMeli,

Things that I have changed so far (f39b577):
- add a docstring to ``pythag`` function
- fix definition of ``unit`` in ``getRodriguezRotation``
- remove unnecessary ``getRotationMatrix``
- replace ``getDist`` by SciPy's ``distance.euclidean``

I furthermore want to get rid of ``dstmev``, which will be covered in the next commit.
