### 2021-03-17 AstraZeneca/kallisto#23 @RMeli
Great @f3rmion . Given the two problems identified above, I think the `rmsd.py` module could do with some dedicated tests.
