### 2021-03-17 dba8a28
Remove dstmev, givens4, pythag from rmsd (#23).
