+++ src/kallisto/rmsd.py
@@ -4,6 +4,7 @@ from collections import Counter
 from typing import Tuple
 
 import numpy as np
+from scipy.sparse.linalg.eigen.arpack import eigsh
 from scipy.spatial import distance
 from scipy.spatial.transform import Rotation as R
 
@@ -53,6 +54,7 @@ def rmsd(n: int, coord1: np.ndarray, coord2: np.ndarray) -> Tuple[float, np.ndar
             rmat[i, j] = np.dot(x[:, i], y[:, j])
 
     # calculate the S matrix (quaternions)
+    # use fac = -1 instead of u.T below
     fac = -1
     smat = np.zeros(shape=(4, 4), dtype=np.float64)
     smat[0, 0] = rmat[0, 0] + rmat[1, 1] + rmat[2, 2]
@@ -76,7 +78,8 @@ def rmsd(n: int, coord1: np.ndarray, coord2: np.ndarray) -> Tuple[float, np.ndar
     smat[3, 3] = -rmat[0, 0] - rmat[1, 1] + rmat[2, 2]
 
     # calculate largest eigenvalue and eigenvector
-    eigenval, eigenvec = dstmev(smat)
+    eigenval, eigenvec = eigsh(smat, 1, which="LA")
+    eigenvec = eigenvec.squeeze()
 
     # convert quaternion eigenvec to rotation matrix U
     u = rotationMatrix(eigenvec)
@@ -128,164 +131,6 @@ def rotationMatrix(q: np.ndarray) -> np.ndarray:
     u[2, 2] = q00 + q33
 
     return u
-
-
-def dstmev(smat: np.ndarray):
-    """Compute the leading eigenvalue and eigenvector of a symmetric,
-    traceless 4x4 matrix A by an inverse power iteration:
-    1) convert matrix to tridiagonal form by rotations
-     V*A*V' = T
-    2) use Gershgorin's theorem to estimate a lower bound
-    for the leading negative eigenvalue:
-     lambda > g=min(T00-t01,-t10+T11-t12,-t21+T22-t23,-t23+T33)
-     where tij=abs(Tij)
-    3) create positive definite matrix
-     B = T-gI
-    4) apply single value decomposition to compute eigenvalues and
-    eigenvectors for SPD matrix B
-    5) shift the spectrum back and keep leading singular vector and largest
-    eigenvalue
-    6) convert eigenvalue to original matrix A by multiplication with V'"""
-
-    # convert triogonal form
-    tmat, vmat = givens4(smat)
-
-    # 2) estimate lower bond of smallest eigenvalue by Gershgorin's theorem
-    eigenval = np.amin(
-        [
-            tmat[0, 0] - abs(tmat[0, 1]),
-            -abs(tmat[1, 0]) + tmat[1, 1] - abs(tmat[1, 2]),
-            -abs(tmat[2, 1]) + tmat[2, 2] - abs(tmat[2, 3]),
-            -abs(tmat[3, 2]) + tmat[3, 3],
-        ]
-    )
-
-    # 3) form positive definite matrix
-    #    T = lambda * I - T
-    for i in range(4):
-        tmat[i, i] = tmat[i, i] - eigenval
-
-    # 4) compute singular values/vectors of SPD matrix
-    from scipy import linalg
-
-    U, s, Vh = linalg.svd(tmat)
-
[…]
