### 2021-03-17 AstraZeneca/kallisto#23 @f3rmion
Hi @RMeli,

I removed ``dstmev``, ``givens4``, and ``pythag`` and replaced them with SciPy analoga.

> Great @f3rmion . Given the two problems identified above, I think the rmsd.py module could do with some dedicated tests.

Do you mean to increase the unit tests to a certain amount?
