### 2021-03-18 47d628d
Add unit tests for correct exs and structure rotation. Close #23
