+++ tests/test_cli.py
@@ -745,6 +745,225 @@ def test_cli_exs(runner):
             os.remove(constrain)
 
 
+def test_cli_exs_works_on_structures(runner):
+    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8", suffix=".xyz") as f1:
+        f1.write("23" + s)
+        f1.write(s)
+        f1.write("C  1.02585 -1.41819  0.10735" + s)
+        f1.write("C  2.19641 -0.65325  0.04371" + s)
+        f1.write("C  2.14922  0.73805 -0.07490" + s)
+        f1.write("C  0.92303  1.41303 -0.14372" + s)
+        f1.write("C -0.26080  0.66414 -0.08355" + s)
+        f1.write("N -1.57019  1.07797 -0.13272" + s)
+        f1.write("C -2.29213 -0.01844 -0.01109" + s)
+        f1.write("C -3.77721 -0.06325  0.01110" + s)
+        f1.write("C -4.61835 -1.18453  0.21920" + s)
+        f1.write("C -5.92028 -0.73409  0.16230" + s)
+        f1.write("N -5.89155  0.61525 -0.08364" + s)
+        f1.write("C -4.59004  1.03318 -0.17187" + s)
+        f1.write("N -1.48774 -1.12334  0.09802" + s)
+        f1.write("C -0.18033 -0.71825  0.04754" + s)
+        f1.write("H  1.06181 -2.49771  0.20128" + s)
+        f1.write("H  3.16184 -1.15317  0.08951" + s)
+        f1.write("H  3.07465  1.30761 -0.11522" + s)
+        f1.write("H  0.87864  2.49164 -0.23712" + s)
+        f1.write("H -4.32389 -2.20831  0.40453" + s)
+        f1.write("H -6.86176 -1.25464  0.27676" + s)
+        f1.write("H -6.70335  1.21181 -0.18081" + s)
+        f1.write("H -4.34671  2.07070 -0.36139" + s)
+        f1.write("H -1.81331 -2.07200  0.19487" + s)
+        f1.flush()
+        f2 = tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8", suffix=".xyz")
+        f2.write("4" + s)
+        f2.write(s)
+        f2.write("C -1.10  0.00  0.00" + s)
+        f2.write("H -1.47  0.73  0.73" + s)
+        f2.write("H -1.47  0.27 -1.00" + s)
+        f2.write("H -1.47 -1.00  0.27" + s)
+        f2.flush()
+        newstructure = "newstructure.xyz"
+        gotFile = os.path.isfile(newstructure)
+        assert gotFile is False
+        result = runner.invoke(
+            cli, ["exs", "--inp", f1.name, f2.name, "--center", "10", "--subnr", "2"],
+        )
+        assert result.exit_code == 0
+        # check Carbon position
+        lookup = "C     -6.7034    1.2118   -0.1808"
+        got = 0
+        want = 25
+        with open(newstructure) as tmpFile:
+            for num, line in enumerate(tmpFile, 1):
+                if lookup in line:
+                    got = num
+        assert got != 0
+        assert got == want
+        gotFile = os.path.isfile(newstructure)
+        assert gotFile is True
+        if gotFile:
+            os.remove(newstructure)
+        constrain = "constrain.inp"
+        gotFile = os.path.isfile(constrain)
+        assert gotFile is True
+        if gotFile:
+            os.remove(constrain)
+
+
+def test_cli_exs_with_rotation(runner):
+    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8", suffix=".xyz") as f1:
+        f1.write("96" + s)
+        f1.write(s)
+        f1.write("N    -1.3672999   -1.4398999    0.1359000" + s)
[…]
