### 2021-03-18 4e03cd5
Prepare release v.1.0.6
