+++ tests/test_version.py
@@ -4,4 +4,4 @@ from kallisto import __version__
 
 
 def test_version():
-    assert __version__ == "1.0.5"
+    assert __version__ == "1.0.6"
