### 2021-03-19 AstraZeneca/kallisto#24 @RMeli
Even better for `getCoordinationNumberSpheres()`, you could compute `cnsp2` and `cnsp3` within the same loop.
