### 2021-03-19 29ab4d7
Exchange option CNSP with PROX and enable size input for shells. Remove code duplication (#24)
