### 2021-03-19 AstraZeneca/kallisto#24 @f3rmion
Hi @RMeli,

for a long time, I was also not happy with the name of the ``CNSP``, as the definition is not really a sphere but rather a shell. So I have now updated this feature as "Proximity Shell" (``Prox``) and also added the possibility of user input via the ``size`` flag (default:  2->3). In the [documentation](https://ehjc.gitbook.io/kallisto/features/prox), I have separated the ``CN`` and the ``Prox`` part accordingly. In addition, I have added how the parameters of the ``CN`` have been fitted and cleaned up some sections in the code.
