### 2021-03-19 AstraZeneca/kallisto#24 @RMeli
Less code _and_ more general, that's great! ;)
