### 2021-03-20 dc3f6fb
Add origin and partner index to Sterimol output.
