### 2021-03-21 #3050 @Sulstice
@f3rmion Your welcome! I learned a lot from your paper digging through it and glad you liked the suggestions :) I did not get another invitation. Maybe I can trigger this whedon bot.
