### 2021-03-22 9b89b1f
Add Numpy and Scipy references to paper.md (Close #25)
