+++ paper.bib
@@ -210,3 +210,27 @@
   Year={2020},
   DOI={10.1002/wcms.1493},
 }
+
+@article{harris2020,
+  url={https://doi.org/10.1038/s41586-020-2649-2},
+  Title={{Array programming with NumPy}},
+  Author={{Harris}, C. R. and {Millman}, K. J. and {van der Walt}, S. J. and {Gommers}, R. and {Virtanen}, P. and {Cournapeau}, D. and {Wieser}, E. and {Taylor}, J. and {Berg}, S. and {Smith}, N. J. and others},
+  Journal={Nature},
+  Volume={585},
+  Number={7825},
+  Pages={357--362},
+  Year={2020},
+  DOI={10.1038/s41586-020-2649-2},
+}
+
+@article{virtanen2020,
+  url={https://doi.org/10.1038/s41592-019-0686-2},
+  Title={{SciPy 1.0: fundamental algorithms for scientific computing in Python}},
+  Author={{Virtanen}, P. and {Gommers}, R. and {Oliphant}, T. E. and {Haberland}, M. and {Reddy}, T. and {Cournapeau}, D. and {Burovski}, E. and {Peterson}, P. and {Weckesser}, W. and {Bright}, J. and others},
+  Journal={Nat. Methods},
+  Volume={17},
+  Number={3},
+  Pages={261--272},
+  Year={2020},
+  DOI={10.1038/s41592-019-0686-2},
+}
