+++ src/kallisto/grid.py
@@ -29,11 +29,56 @@ gridSize = dict(
         (8, 146),
         (9, 170),
         (10, 194),
+        (11, 230),
+        (12, 266),
+        (13, 302),
+        (14, 350),
+        (15, 434),
+        (16, 590),
+        (17, 770),
+        (18, 974),
+        (19, 1202),
+        (20, 1454),
+        (21, 1730),
+        (22, 2030),
     ]
 )
 
 # keep octahedral symmetry
 z = 0.0
+
+
+def getGridFromInt(n: int, grid: np.ndarray, weights: np.ndarray, count: Counter):
+    """Returns Lebedev-Laikov grid, weights, and count depending on requested
+    gridSize integer."""
+
+    switcher = {
+        0: lambda: ld0006(grid, weights, count),
+        1: lambda: ld0014(grid, weights, count),
+        2: lambda: ld0026(grid, weights, count),
+        3: lambda: ld0038(grid, weights, count),
+        4: lambda: ld0050(grid, weights, count),
+        5: lambda: ld0074(grid, weights, count),
+        6: lambda: ld0086(grid, weights, count),
+        7: lambda: ld0110(grid, weights, count),
+        8: lambda: ld0146(grid, weights, count),
+        9: lambda: ld0170(grid, weights, count),
+        10: lambda: ld0194(grid, weights, count),
+        11: lambda: ld0230(grid, weights, count),
+        12: lambda: ld0266(grid, weights, count),
+        13: lambda: ld0302(grid, weights, count),
+        14: lambda: ld0350(grid, weights, count),
+        15: lambda: ld0434(grid, weights, count),
+        16: lambda: ld0590(grid, weights, count),
+        17: lambda: ld0770(grid, weights, count),
+        18: lambda: ld0974(grid, weights, count),
+        19: lambda: ld1202(grid, weights, count),
+        20: lambda: ld1454(grid, weights, count),
+        21: lambda: ld1730(grid, weights, count),
+        22: lambda: ld2030(grid, weights, count),
+    }  # type: ignore
+    func = switcher.get(n, lambda: "Invalid Lebedev-Laikov grid requested.")
+    return func()
 
 
 def getLebedevLaikovGrid(n: int):
@@ -52,28 +97,8 @@ def getLebedevLaikovGrid(n: int):
     grid = np.zeros(shape=(size, 3), dtype=np.float64)
     weights = np.zeros(shape=(size,), dtype=np.float64)
 
-    if n == 0:
-        grid, weights, count = ld0006(grid, weights, count)
-    elif n == 1:
-        grid, weights, count = ld0014(grid, weights, count)
-    elif n == 2:
-        grid, weights, count = ld0026(grid, weights, count)
-    elif n == 3:
-        grid, weights, count = ld0038(grid, weights, count)
-    elif n == 4:
-        grid, weights, count = ld0050(grid, weights, count)
-    elif n == 5:
-        grid, weights, count = ld0074(grid, weights, count)
-    elif n == 6:
-        grid, weights, count = ld0086(grid, weights, count)
-    elif n == 7:
-        grid, weights, count = ld0110(grid, weights, count)
-    elif n == 8:
-        grid, weights, count = ld0146(grid, weights, count)
-    elif n == 9:
-        grid, weights, count = ld0170(grid, weights, count)
[…]
