### 2021-03-22 #3050 @f3rmion
@Sulstice I added references for Numpy and Scipy in ``paper.md`` and ``paper.bib``.
