### 2021-03-24 293503b
Refactor parts of grid.py
