### 2021-03-24 AstraZeneca/kallisto#26 @RMeli
[JOSS] Partial Charges
## Documentation

In the documentation, I think the variable `chi_i` should be defined after the first formula for `E_IES` instead than within the definition of the matrix elements of `A`.

## Implementation

### Vectorisation

There are many places where the code can be vectorised in `getAtomicPartialCharges()`. This will speedup the execution since it will get rid of multiple `for` loops that grow linearly with the number of atoms. Some places that could benefit from vectorisation are the following:

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L173-L177

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L202-L205

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L207-L210

### Reshaping

I struggle to see why the arrays `A` and `X` are being reshaped

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L215-L216

since they have been initialized with the correct shape already. Am I missing something?

### Consistency with Documentation

Is the following line of code correct?

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L191

If yes, I think a few things would help understanding what is going on:

1. The name `gam` (which recalls to "gamma") is very similar to `gamij`; if I understand correctly the former represent `Jii` in the documentation and the latter is clearly `gammaij`. I'd make the variable names in the code more consistent with the documentation (or vice-versa)
2. I'd give explicitly give the definition of `gammaij` and `gammaii` in terms of the parameters `alphai` and `alphaj`; from the code there is `gamij = 1.0 / np.sqrt(alpha[i] + alpha[j])`, but then the line above seems to indicate that `gammaii = 1.0 / np.sqrt(alpha[i])` and not `gamii = 1.0 / np.sqrt(alpha[i] + alpha[i])` .

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
