### 2021-03-24 AstraZeneca/kallisto#27 @RMeli
[JOSS] Polarizabilities
## Documentation

> Hence the problem is shifted from calculating dynamic polarizabilities to calculating atomic coordination numbers and interpolate the atomic reference systems. This procedure comes along with a tremendous increase in efficiency and overcomes furthermore numerical issues that could occur in TD-DFT calculations.

- [x] Can you please add a comment on the accuracy of this method at this stage and link to the table?

> Note that the dynamic reference polarizabilities can be pre-scaled to, e.g., incorporate atomic-charge information as has been introduced recently by Caldeweyher et al.

- [x] Add missing reference to "Caldeweyher et al."

Other comments:
- [x] Add timing details to the table? How `kallisto` compares to `AlphaML` in terms of speed?
- [x] Add example with the `--molecular` option to compute the molecular polarizability

## Implementation

*The code is quite complex, it could do with a few more comments so that it easier to follow and understand.*

As mentioned in #26 (and for the reason above) , it would be helpful to have a better consistency between variable names in the code and the documentation. For example, you have `wf = 6.0` in the code, `beta` in the documentation (`beta` is not defined at all in the documentation) and `beta2` in the [paper](https://s3-eu-west-1.amazonaws.com/itempdf74155353254prod/7430216/A_Generally_Applicable_Atomic-Charge_Dependent_London_Dispersion_Correction_Scheme_v2.pdf). I had to dig into the paper to understand what `wf = 6.0` was.

### Vectorisation

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L244-L245

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L314-L315

There might be other parts where the code could benefit from vectorisation and some parts where loops can be reduced.

For example, here I think you can combine the two loops together:

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L290-L299

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
