### 2021-03-24 AstraZeneca/kallisto#26 @awvwgk
> Is the following line of code correct?

Certainly correct but obscured by the fact that unrelated constants are collected in `sqrt2pi`. Most notably the 1/sqrt(2) from the evaluation of γii is merged with the 2/sqrt(π) from the paper / docs.
