### 2021-03-24 AstraZeneca/kallisto#28 @RMeli
[JOSS] Van der Waals Radii
## Documentation

The following docstring is incorrect:

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L383-L390

## Implementation

### Vectorisation

Unfortunately I think that changes introduced for https://github.com/AstraZeneca/kallisto/issues/22 prevent vectorisation in this case (sorry for not spotting this previously!), since you have to go from the atomic number to the chemical symbol in order to extract the parameters needed.

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L406-L408

I'm not too keen to suggest to revert the changes, so it is up to you to decide if it is worth the effort.

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
