### 2021-03-24 AstraZeneca/kallisto#26 @RMeli
@awvwgk thank for confirming this. In such case I think the definition of `Aii` in the documentation (and the paper?) should read `Jii + sqrt(2/π) / sqrt(alphaii)` and not `Jii + sqrt(2/π)  gammaii`.
