### 2021-03-24 AstraZeneca/kallisto#26 @awvwgk
Not quite, the paper states `Aii = Jii + 2/sqrt(π)·γii` which is equivalent to `Aii = Jii + sqrt(2/π) / αi`, special attention must be paid to the fact that the array `alpha` is actually α², see

https://github.com/AstraZeneca/kallisto/blob/293503b9d2b12ef3551fc15f50f50812bc7c2c9d/src/kallisto/methods.py#L177
