### 2021-03-24 AstraZeneca/kallisto#26 @awvwgk
I see, the formula 12 in the preprint is in error here, but the formula in the corresponding JCP paper is given correct.
