### 2021-03-24 AstraZeneca/kallisto#26 @RMeli
Thanks for pointing this out, I was indeed looking at the preprint! The [documentation](https://ehjc.gitbook.io/kallisto/features/eeq) seems to contain the same error, that's why I was confused I guess.

Given that the preprint is linked in the documentation (which is great, since it is accessible to eveybody!) I think a note about this would be very helpful.
