### 2021-03-25 AstraZeneca/kallisto#23 @RMeli
@f3rmion I just noticed you used `scipy.spatial.transform.Rotation` in `getRodriguezRotation()` but not in `rotationMatrix()`. Is this on purpose or just an oversight?
