### 2021-03-25 AstraZeneca/kallisto#26 @f3rmion
Hi @RMeli, Hi @awvwgk,

> I see, the formula 12 in the preprint is in error here, but the formula in the corresponding JCP paper is given correctly.

Yes, there is an error in the preprint since it should either be ``2 Jii / sqrt(π)`` or by using ``Jii = 1 / sqrt( ai² + ai²)`` to get ``sqrt(2) / ( sqrt(π) sqrt(ai²) )``. I'll add a note to the docs.

> The documentation seems to contain the same error, that's why I was confused I guess.

The docs seem to be right to me, where do you see this error? Or do you mean according to the preprint? :)
