### 2021-03-25 AstraZeneca/kallisto#26 @awvwgk
The formula is present twice in the documentation, the one in the energy expression is correct, the A matrix elements are in error.
