### 2021-03-25 AstraZeneca/kallisto#26 @RMeli
@f3rmion I think the `Aij` coefficient in the documentation for `i = j` should read `Jii + 2 / sqrt(π) * gammaii` instead of `Jii + sqrt(2/π) * gammaii`?
