### 2021-03-25 e3c9ef2
Vectorize for loops in eeq and remove reshape (#26)
