### 2021-03-25 AstraZeneca/kallisto#26 @f3rmion
Hi @RMeli,

in e3c9ef2 all mentioned ``for``-loops have been vectorized and the useless reshape has been deleted. I furthermore updated the docs and added a note regarding the erroneous eq. 12 in the linked preprint. Thanks for this nice catch!
