### 2021-03-25 AstraZeneca/kallisto#27 @awvwgk
> * Add timing details to the table? How `kallisto` compares to `AlphaML` in terms of speed?

I would be intrigued to see how the performance of `kallisto` and the reference implementation in [`dftd4`](https://github.com/dftd4/dftd4/tree/main/python) differ in terms of speed as well in such a comparison.
