### 2021-03-25 fe84963
Remove regularization term from eeq (#26)
