### 2021-03-25 AstraZeneca/kallisto#27 @f3rmion
Hi @RMeli,

> Can you please add a comment on the accuracy of this method at this stage and link to the table?

I added a comment and linked to the table, respectively.

>  Add missing reference to "Caldeweyher et al."

Another link has been added :)

>  Add timing details to the table? How kallisto compares to AlphaML in terms of speed?

I created a small benchmark set using structures from [this repository](https://github.com/f3rmion/molpol135) and calculated timings for both methods. For the ``AlphaML`` one I used the value they print out in their webinterface to get timings (CCSD mode). Here, they print out the timing to compute the quantities. I, therefore, did the same thing for ``kallisto``, but I averaged over three different runs for every molecule and used the averaged value in the final summation to obtain total computational timings.

> Add example with the --molecular option to compute the molecular polarizability

Another example is given for molecular polarizabilities using the Alanine-Glycine molecule.
