### 2021-03-26 AstraZeneca/kallisto#27 @RMeli
Nice!

> This is definitely a comparison that I'll add to the roadmap 2021, however, I'm already quite sure that the implementation with pre-compiled dftd4 library can't be beaten in terms of efficiency. ;)

That's probably true, although it would be great to see how close one can get using `numpy`/`scipy`.
