### 2021-03-26 AstraZeneca/kallisto#27 @awvwgk
It's almost an order of magnitude for me, (Skylake, Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz). Conda-forge's Python, Scipy and MKL (average over 5 runs for the 45 systems from the docs):

1: `kallisto` CLI -> 15.4s user, 9.1s system
2: `dftd4` CLI -> 2.3s user, 0.4s system
3: `dftd4` Python API > dftd4.out -> 0.4s user, 0.2s

Looks like the task in docs could be bound by launching the processes.
