### 2021-03-26 AstraZeneca/kallisto#27 @f3rmion
Thanks, @awvwgk! I'll update the docs with your timings then :)
