### 2021-03-27 AstraZeneca/kallisto#30 @RMeli
[JOSS] Sterimol Descriptors
# Documentation

The documentation for the sterimol descriptors is quite scarce. I'd add some details (or at very least link to the relevant section of the [Morfeus package](https://kjelljorner.github.io/morfeus/sterimol.html).

# Code

Do I understand correctly that the code comes directly from [Morfeus package](https://kjelljorner.github.io/morfeus/sterimol.html)? I would make this clearer in the module docstring (not only the function `getClassicalSterimol()`). Is there any kind of copyright or license that needs to be mentioned as well?

I was not able to find this code; if it was given in private, maybe an acknowledgment of this in the paper is needed. Some clarifications about how the code was obtained and integrated would be helpful to better understand the situation.

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
