### 2021-03-28 AstraZeneca/kallisto#31 @RMeli
[JOSS] Tests
Most of the tests are in `test_cli.py` instead of being in separate and standalone files (for example `test_methods.py`, `test_rmsd.py`, ...). This is not a problem in itself, but it seems that in some tests the `assert`s do not really check the result.

Using the CLI to perform tests forces you to check the results as strings instead of numbers. This seems to be done incorrectly (see below). Additionally, if at some point you want to change the output format (for example to use scientific notation), you will have to change all the assertions.

I would suggest to re-work the tests in order to decouple the test of the CLI from the tests of the functionality. This is not really necessary, but the feeling I have looking at `test_cli.py` is that the code coverage is very high because all parts of the code are run while some tests are not really meaningful (i.e. they mostly check that the test was run and output files have been created). I can't see many tests for edge cases either.  I would like to see more clear and meaningful tests, comparing with known results (obtained from reference software or from manual calculations), so that users can make sure the package produces correct numerical results.

# `testcli.py`

## Tests without output?

The following tests does not seem to have any output when run outside the testsuite:

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L96-L104

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L107-L115

## Test checking only if output files have been produced

The following tests only check if files have been produced, but not the content of the output file and there is no check beside the fact that they run:

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L118-L133

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L418-L433

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L730-L745

## Incorrect assertions

The following assertions look incorrect, since `"A" and "B"` evaluates to `"B"` if `"A"` is true (i.e. non-empty string):

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L202

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L214

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L226

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L250

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L262

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L274

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L310

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L322

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L336

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L350

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L364

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L378

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L459

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L471

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L1182-L1186

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L1236

# Additional Comments

The following test brought to my attention that the original information in the second line of the XYZ file is lots and replaced by "Created with kallisto":

https://github.com/AstraZeneca/kallisto/blob/fe8496349b48f80c2b5af41e9754b87aef932569/tests/test_cli.py#L175-L178

I think it would be better to retain the original comment and add "Created with kallisto" after it.

---

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/3050)
