### 2021-03-28 AstraZeneca/kallisto#31 @RMeli
Maybe it is worth merging `test_xyz.py` into `test_strucreader.py` since `strucreader.py` deals with both XYZ and Turbomole input?
