### 2021-03-28 AstraZeneca/kallisto#27 @RMeli
@f3rmion LGTM. One last thing; could you please move the following

> MAD = Mean Absolute Deviation; MD = Mean Deviation; RMSD = Root-Mean Squared Deviation; AMAX = Absolut MAXimum deviation.

just after the first table? I think it will help to better separate the two tables and avoid any confusion (`dftd4` being aligned with the MP2 column).
