### 2021-03-28 AstraZeneca/kallisto#27 @f3rmion
Hi @RMeli,

> just after the first table? I think it will help to better separate the two tables and avoid any confusion (dftd4 being aligned with the MP2 column).

Done :)
