### 2021-03-28 AstraZeneca/kallisto#30 @f3rmion
Hi @RMeli,

> The documentation for the sterimol descriptors is quite scarce. I'd add some details (or at very least link to the relevant section of the Morfeus package.

I will expand the documentation accordingly so that the main aspect of the Sterimol descriptors becomes clear.

> Do I understand correctly that the code comes directly from Morfeus package? I would make this clearer in the module docstring (not only the function getClassicalSterimol()). Is there any kind of copyright or license that needs to be mentioned as well?
> I was not able to find this code; if it was given in private, maybe an acknowledgment of this in the paper is needed. Some clarifications about how the code was obtained and integrated would be helpful to better understand the situation.

The Morfeus code is not yet published, which will change in the near future (open source). The code is not directly taken from this package, however, the algorithm has been adapted (therefore the mentioning). When I asked him the last time he only wanted a note in the docs, but I'll definitely add an acknowledgment to the paper ;)
