### 2021-03-29 230d91b
Add input as positional argument and change output to be an option.
