+++ tests/test_cli.py
@@ -39,9 +39,7 @@ def test_cli_cns_exp_silent(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(
-            cli, ["--silent", "cns", "--inp", f.name, "--cntype", "exp"]
-        )
+        result = runner.invoke(cli, ["--silent", "cns", "--cntype", "exp", f.name])
         assert result.exit_code == 0
 
 
@@ -52,7 +50,7 @@ def test_cli_cns_exp(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["cns", "--inp", f.name, "--cntype", "exp"])
+        result = runner.invoke(cli, ["cns", "--cntype", "exp", f.name])
         assert result.exit_code == 0
         assert "0.98678927" in result.output
 
@@ -64,7 +62,7 @@ def test_cli_cns_cov(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["cns", "--inp", f.name, "--cntype", "cov"])
+        result = runner.invoke(cli, ["cns", "--cntype", "cov", f.name])
         assert result.exit_code == 0
         assert "0.9189476" in result.output
 
@@ -76,7 +74,7 @@ def test_cli_cns_erf(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["cns", "--inp", f.name, "--cntype", "erf"])
+        result = runner.invoke(cli, ["cns", "--cntype", "erf", f.name])
         assert result.exit_code == 0
         assert "0.9878465" in result.output
 
@@ -89,7 +87,7 @@ def test_cli_bonds_silent(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["--silent", "bonds", "--inp", f.name])
+        result = runner.invoke(cli, ["--silent", "bonds", f.name])
         assert result.exit_code == 0
 
 
@@ -100,7 +98,7 @@ def test_cli_bonds(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["bonds", "--inp", f.name])
+        result = runner.invoke(cli, ["bonds", f.name])
         assert result.exit_code == 0
 
 
@@ -111,7 +109,7 @@ def test_cli_bonds_partner(runner):
         f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
         f.write("$end")
         f.flush()
-        result = runner.invoke(cli, ["bonds", "--inp", f.name, "--partner", "0"])
+        result = runner.invoke(cli, ["bonds", "--partner", "0", f.name])
         assert result.exit_code == 0
 
 
@@ -125,7 +123,7 @@ def test_cli_bonds_constrain(runner):
         constrain = "constrain.inp"
         gotFile = os.path.isfile(constrain)
         assert gotFile is False
-        result = runner.invoke(cli, ["bonds", "--inp", f.name, "--constrain"])
+        result = runner.invoke(cli, ["bonds", "--constrain", f.name])
         assert result.exit_code == 0
         gotFile = os.path.isfile(constrain)
         assert gotFile is True
@@ -150,7 +148,7 @@ def test_cli_sort(runner):
[…]
