+++ src/kallisto/console.py
@@ -28,7 +28,7 @@ pass_config = click.make_pass_decorator(Config, ensure=True)
 def cli(config, silent: bool, shift: int):
     """kallisto calculates quantum mechanically derived atomic features.\n
 
-    Please check out the documentation.
+    Please check out the documentation (https://ehjc.gitbook.io/kallisto/).
     """
 
     config.shift = shift
@@ -38,21 +38,21 @@ def cli(config, silent: bool, shift: int):
 @cli.command("cns")
 @pass_config
 @click.option(
-    "--inp",
-    default="coord",
-    type=str,
-    show_default=True,
-    required=True,
-    help="Geometry file in xyz or turbomole format.",
-)
-@click.option(
     "--cntype",
     default="erf",
     type=str,
     show_default=True,
     help="Coordination number type (exp, cov, err).",
 )
-@click.argument("out", type=click.File("w"), default="-", required=False)
+@click.option(
+    "--out",
+    default="-",
+    type=click.File("w"),
+    show_default=True,
+    required=False,
+    help="Write to output file.",
+)
+@click.argument("inp", type=str, default="coord", required=True)
 def cns(config, inp: str, out: click.File, cntype: str):
     """Atomic coordination numbers."""
 
@@ -68,14 +68,6 @@ def cns(config, inp: str, out: click.File, cntype: str):
 @cli.command("prox")
 @pass_config
 @click.option(
-    "--inp",
-    default="coord",
-    type=str,
-    show_default=True,
-    required=True,
-    help="Geometry file in xyz or turbomole format.",
-)
-@click.option(
     "--size",
     type=(int, int),
     default=(2, 3),
@@ -83,7 +75,15 @@ def cns(config, inp: str, out: click.File, cntype: str):
     required=False,
     help="Size that defines the proximity shells.",
 )
-@click.argument("out", type=click.File("w"), default="-", required=False)
+@click.option(
+    "--out",
+    default="-",
+    type=click.File("w"),
+    show_default=True,
+    required=False,
+    help="Write to output file.",
+)
+@click.argument("inp", type=str, default="coord", required=True)
 def prox(config, inp: str, size: Tuple[int, int], out: click.File):
     """Atomic proximity shells."""
 
@@ -99,14 +99,6 @@ def prox(config, inp: str, size: Tuple[int, int], out: click.File):
 @cli.command("bonds")
 @pass_config
 @click.option(
-    "--inp",
-    default="coord",
-    type=str,
-    show_default=True,
-    required=True,
-    help="Geometry file in xyz or turbomole format.",
-)
-@click.option(
     "--partner",
     type=str,
     default="X",
@@ -115,7 +107,15 @@ def prox(config, inp: str, size: Tuple[int, int], out: click.File):
     help="Get all partner of atom X.",
 )
 @click.option("--constrain", is_flag=True)
-@click.argument("out", type=click.File("w"), default="-", required=False)
+@click.option(
+    "--out",
+    default="-",
+    type=click.File("w"),
+    show_default=True,
+    required=False,
+    help="Write to output file.",
+)
+@click.argument("inp", type=str, default="coord", required=True)
 def bonds(config, inp: str, partner: int, constrain: bool, out: click.File):
     """Get information about covalent bonding partner."""
 
@@ -153,14 +153,6 @@ def bonds(config, inp: str, partner: int, constrain: bool, out: click.File):
 @cli.command("sort")
 @pass_config
 @click.option(
-    "--inp",
-    default="coord",
-    type=str,
-    show_default=True,
-    required=True,
-    help="Geometry file in xyz or turbomole format.",
-)
-@click.option(
     "--start",
     type=str,
     default="X",
[…]

[…]
