### 2021-03-29 934000a
Correct docstring for getVanDerWaalsRadii (#28)
