### 2021-03-29 AstraZeneca/kallisto#28 @f3rmion
Hi @RMeli,

> The following docstring is incorrect:

I updated the docstring to get rid of the copy-paster error.

> I'm not too keen to suggest to revert the changes, so it is up to you to decide if it is worth the effort.

Yes, at the moment I don't see that it's worth the effort.
