### 2021-03-30 AstraZeneca/kallisto#15 @RMeli
@f3rmion In the manuscript, I'd specify that the Van der Waals radii are environment and charge dependent in the list under "Atomic Features".
