### 2021-03-31 c0c6ce4
Use SciPy rotation in rotationMatrix in rmsd (#23)
