### 2021-03-31 AstraZeneca/kallisto#30 @f3rmion
Hi @RMeli,

I added an acknowledgment to the paper (17429a2) and extended the [docs](https://ehjc.gitbook.io/kallisto/features/stm) with an explanation of what Sterimol descriptors are.
