+++ noxfile.py
@@ -10,7 +10,9 @@ nox.options.sessions = "lint", "mypy", "pytype", "tests"
 locations = "src", "tests", "noxfile.py"
 
 
-@nox.session(python=["3.8", "3.7"])
+python_versions = ["3.9, "3.8", "3.7"]
+
+@nox.session(python=python_versions)
 def tests(session: Session) -> None:
     args = session.posargs or ["--cov", "-m", "not e2e"]
     session.run("poetry", "install", "--no-dev", external=True)
@@ -20,7 +22,7 @@ def tests(session: Session) -> None:
     session.run("pytest", *args)
 
 
-@nox.session(python=["3.8", "3.7"])
+@nox.session(python=python_versions)
 def lint(session: Session) -> None:
     args = session.posargs or locations
     install_with_constraints(
@@ -43,7 +45,7 @@ def black(session: Session) -> None:
 
 @nox.session(python="3.8")
 def safety(session: Session) -> None:
-    with tempfile.NamedTemporaryFile() as requirements:
+    with tempfile.NamedTemporaryFile(dir="./tmp") as requirements:
         session.run(
             "poetry",
             "export",
@@ -57,7 +59,7 @@ def safety(session: Session) -> None:
         session.run("safety", "check", f"--file={requirements.name}", "--full-report")
 
 
-@nox.session(python=["3.8", "3.7"])
+@nox.session(python=python_versions)
 def mypy(session: Session) -> None:
     args = session.posargs or locations
     install_with_constraints(session, "mypy")
@@ -81,7 +83,7 @@ def coverage(session: Session) -> None:
 
 
 def install_with_constraints(session: Session, *args, **kwargs) -> None:
-    with tempfile.NamedTemporaryFile() as requirements:
+    with tempfile.NamedTemporaryFile(dir="./tmp") as requirements:
         session.run(
             "poetry",
             "export",
