+++ noxfile.py
@@ -10,7 +10,7 @@ nox.options.sessions = "lint", "mypy", "pytype", "tests"
 locations = "src", "tests", "noxfile.py"
 
 
-python_versions = ["3.9, "3.8", "3.7"]
+python_versions = ["3.9", "3.8", "3.7"]
 
 @nox.session(python=python_versions)
 def tests(session: Session) -> None:
