+++ noxfile.py
@@ -13,12 +13,16 @@ locations = "src", "tests", "noxfile.py"
 
 python_versions = ["3.9", "3.8", "3.7"]
 
+
 @nox.session(python=python_versions)
 def tests(session: Session) -> None:
     args = session.posargs or ["--cov", "-m", "not e2e"]
     session.run("poetry", "install", "--no-dev", external=True)
     install_with_constraints(
-        session, "coverage[toml]", "pytest", "pytest-cov",
+        session,
+        "coverage[toml]",
+        "pytest",
+        "pytest-cov",
     )
     session.run("pytest", *args)
 
