+++ noxfile.py
@@ -19,10 +19,7 @@ def tests(session: Session) -> None:
     args = session.posargs or ["--cov", "-m", "not e2e"]
     session.run("poetry", "install", "--no-dev", external=True)
     install_with_constraints(
-        session,
-        "coverage[toml]",
-        "pytest",
-        "pytest-cov",
+        session, "coverage[toml]", "pytest", "pytest-cov",
     )
     session.run("pytest", *args)
 
