### 2021-04-01 AstraZeneca/kallisto#31 @RMeli
Most tests seems to re-define the same file. Probably it is worth using [fixtures](https://docs.pytest.org/en/stable/tmpdir.html#the-tmpdir-factory-fixture)?

This should also get rid of the problems in #32 (hopefully), if you are keen to support all platforms.
