### 2021-04-02 AstraZeneca/kallisto#31 @f3rmion
> Most tests seems to re-define the same file. Probably it is worth using [fixtures](https://docs.pytest.org/en/stable/tmpdir.html#the-tmpdir-factory-fixture)?
>
> This should also get rid of the problems in #32 (hopefully), if you are keen to support all platforms.

At the moment I am working on separating all the tests regarding the functionality of the CLI and numerical tests. So soon there will be no more string comparisons. When this is finished, I will take care of providing the CLI unit tests (which need files) with a corresponding fixture, so that Windows, among others, also gets the corresponding support.
