### 2021-04-02 5bd6298
Separate CLI functionality unit tests from numerical ones (#31)
