+++ tests/test_cli.py
@@ -1,7 +1,6 @@
 # tests/test_cli.py
 
 import os
-import tempfile
 
 import click.testing
 import pytest
@@ -32,1141 +31,307 @@ def test_cli_print_help(runner):
 
 
 # test cli part for coordination number
-def test_cli_cns_exp_silent(runner):
-    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
-        f.write("$coord" + s)
-        f.write("0.00000000000000 0.00000000000000 -1.06176434496059 c" + s)
-        f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
-        f.write("$end")
-        f.flush()
-        result = runner.invoke(cli, ["--silent", "cns", "--cntype", "exp", f.name])
-        assert result.exit_code == 0
+def test_cli_cns_exp_silent(runner, pyridine_xyz):
+    result = runner.invoke(cli, ["--silent", "cns", "--cntype", "exp", pyridine_xyz])
+    assert result.exit_code == 0
 
 
-def test_cli_cns_exp(runner):
-    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
-        f.write("$coord" + s)
-        f.write("0.00000000000000 0.00000000000000 -1.06176434496059 c" + s)
-        f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
-        f.write("$end")
-        f.flush()
-        result = runner.invoke(cli, ["cns", "--cntype", "exp", f.name])
-        assert result.exit_code == 0
-        assert "0.98678927" in result.output
+def test_cli_cns_exp(runner, pyridine_xyz):
+    result = runner.invoke(cli, ["cns", "--cntype", "exp", pyridine_xyz])
+    assert result.exit_code == 0
 
 
-def test_cli_cns_cov(runner):
-    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
-        f.write("$coord" + s)
-        f.write("0.00000000000000 0.00000000000000 -1.06176434496059 c" + s)
-        f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
-        f.write("$end")
-        f.flush()
-        result = runner.invoke(cli, ["cns", "--cntype", "cov", f.name])
-        assert result.exit_code == 0
-        assert "0.9189476" in result.output
+def test_cli_cns_cov(runner, pyridine_xyz):
+    result = runner.invoke(cli, ["cns", "--cntype", "cov", pyridine_xyz])
+    assert result.exit_code == 0
 
 
-def test_cli_cns_erf(runner):
-    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
-        f.write("$coord" + s)
-        f.write("0.00000000000000 0.00000000000000 -1.06176434496059 c" + s)
-        f.write("0.00000000000000 0.00000000000000  1.06176434496059 h" + s)
-        f.write("$end")
-        f.flush()
-        result = runner.invoke(cli, ["cns", "--cntype", "erf", f.name])
-        assert result.exit_code == 0
-        assert "0.9878465" in result.output
+def test_cli_cns_erf(runner, pyridine_xyz):
+    result = runner.invoke(cli, ["cns", "--cntype", "erf", pyridine_xyz])
+    assert result.exit_code == 0
 
 
-def test_cli_cns_invalid(runner):
-    with tempfile.NamedTemporaryFile(mode="w+", encoding="utf-8") as f:
-        f.write("$coord" + s)
-        f.write("0.00000000000000 0.00000000000000 -1.06176434496059 c" + s)
[…]
