### 2021-04-05 40f8ed6
Set correct docstring in conftest, simplify exs test.
