### 2021-04-05 df76e84
Remove pyridine_mH fixture in exs.
