### 2021-04-05 AstraZeneca/kallisto#31 @f3rmion
Hi @RMeli,

I replaced all ``tempfile.NamedTemporaryFile`` with files in temporary directories (using ``tmpdir_factory`` and fixtures) and (after some strange errors that I've fixed) it is now working. You can try to extend the platform support now.
