### 2021-04-06 91154dc
Clean up lgtm alerts.
