### 2021-04-06 AstraZeneca/kallisto#31 @RMeli
@f3rmion There are still a few `tempfile.NamedTemporaryFile` in `test_molecule.py` and `test_structreader.py` as far as I can see. I can take care of them as part of #32 if you want (since they are making the Windows CI fail).
