+++ noxfile.py
@@ -6,7 +6,7 @@ import tempfile
 import nox
 from nox.sessions import Session
 
-nox.options.sessions = "lint", "mypy", "tests"
+nox.options.sessions = "lint", "mypy", "pytype", "tests"
 
 locations = "src", "tests", "noxfile.py"
 
