### 2021-04-07 2e4c364
remove windows from ci matrix
