+++ .github/workflows/tests.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macOS-latest, windows-latest]
+        os: [ubuntu-latest, macOS-latest]
         python-version: ['3.7', '3.8', '3.9']
     name: Python ${{ matrix.python-version }} OS ${{ matrix.os }}
     steps:
