### 2021-04-07 806ef7d
Expand test matrix to macOS (#32)
* expand test matrix

* change directory for tempfile

* fix python versions list

* tempfile in current dir with os module

* fix noxfile linting

* revert change in noxfile

* try to use tempdirs instead of tempfiles

* remove temporary files

* fix linting

* fix linting

* try to remove pytype

* add back pytype

* remove windows from ci matrix

Signed-off-by: Rocco Meli <rocco.meli@biodtp.ox.ac.uk>
