+++ noxfile.py
@@ -1,5 +1,6 @@
 # noxfile.py
 
+import os
 import tempfile
 
 import nox
@@ -10,7 +11,10 @@ nox.options.sessions = "lint", "mypy", "pytype", "tests"
 locations = "src", "tests", "noxfile.py"
 
 
-@nox.session(python=["3.8", "3.7"])
+python_versions = ["3.9", "3.8", "3.7"]
+
+
+@nox.session(python=python_versions)
 def tests(session: Session) -> None:
     args = session.posargs or ["--cov", "-m", "not e2e"]
     session.run("poetry", "install", "--no-dev", external=True)
@@ -20,7 +24,7 @@ def tests(session: Session) -> None:
     session.run("pytest", *args)
 
 
-@nox.session(python=["3.8", "3.7"])
+@nox.session(python=python_versions)
 def lint(session: Session) -> None:
     args = session.posargs or locations
     install_with_constraints(
@@ -43,21 +47,26 @@ def black(session: Session) -> None:
 
 @nox.session(python="3.8")
 def safety(session: Session) -> None:
-    with tempfile.NamedTemporaryFile() as requirements:
+    with tempfile.TemporaryDirectory() as tmpdirname:
         session.run(
             "poetry",
             "export",
             "--dev",
             "--format=requirements.txt",
             "--without-hashes",
-            f"--output={requirements.name}",
+            f"--output={os.path.join(tmpdirname, 'poetry-export.out')}",
             external=True,
         )
         install_with_constraints(session, "safety")
-        session.run("safety", "check", f"--file={requirements.name}", "--full-report")
+        session.run(
+            "safety",
+            "check",
+            f"--file={os.path.join(tmpdirname, 'safety-check.out')}",
+            "--full-report",
+        )
 
 
-@nox.session(python=["3.8", "3.7"])
+@nox.session(python=python_versions)
 def mypy(session: Session) -> None:
     args = session.posargs or locations
     install_with_constraints(session, "mypy")
@@ -81,14 +90,18 @@ def coverage(session: Session) -> None:
 
 
 def install_with_constraints(session: Session, *args, **kwargs) -> None:
-    with tempfile.NamedTemporaryFile() as requirements:
+    with tempfile.TemporaryDirectory() as tmpdirname:
         session.run(
             "poetry",
             "export",
             "--dev",
             "--format=requirements.txt",
             "--without-hashes",
-            f"--output={requirements.name}",
+            f"--output={os.path.join(tmpdirname, 'poetry-export.out')}",
             external=True,
         )
-        session.install(f"--constraint={requirements.name}", *args, **kwargs)
+        session.install(
+            f"--constraint={os.path.join(tmpdirname, 'poetry-export.out')}",
+            *args,
+            **kwargs,
+        )
