### 2021-04-07 AstraZeneca/kallisto#31 @f3rmion
Hi @RMeli,

most of the tests have now explicit checks against numerical values. The CLI tests check only if everything runs smoothly through without any error. I guess the suggestion of using fixtures was really good since the unit tests are cleaner now. I furthermore added some new tests that, e.g., check for implemented options (cns and vdw). The ``turbomole`` and ``xyz`` tests are now in the ``strucreader`` as well. The input of created constrainment files are now checked as well. Do you think this issue can be closed, or is something still missing?
