### 2021-04-07 AstraZeneca/kallisto#31 @RMeli
@f3rmion Thanks for changing the tests to be more explicit and robust! I know it was a lot of work, but I think it was well worth going forward.

For a skim trough they look very good, but I'll have a deeper look in the coming days and update or close this issue. I hope to finish the whole review this weekend at the latest.

One thing I noticed skimming through them is that assertions like the following should have an absolute value:
https://github.com/AstraZeneca/kallisto/blob/58602bd713d8e989b631eee7722514eadd04e9ca/tests/test_eeq.py#L13-L14

Maybe it would be even better to use [`math.isclose`](https://docs.python.org/dev/library/math.html#math.isclose) and [`numpy.isclose`](https://numpy.org/doc/stable/reference/generated/numpy.allclose.html)
