### 2021-04-07 AstraZeneca/kallisto#31 @f3rmion
Good suggestion, I go with ``numpy.isclose`` then.
