### 2021-04-07 ec95179
Use numpy.isclose in unit tests. Fix wrong naming in fixture.
