+++ tests/test_stm.py
@@ -1,10 +1,9 @@
 # tests/test_stm.py
 
+import numpy as np
+
 from kallisto.sterics import getClassicalSterimol
 from tests.store import toluene
-
-# global epsilon
-epsilon = 1e-06
 
 
 def test_stm():
@@ -12,6 +11,6 @@ def test_stm():
     origin = 6
     partner = 5
     L, bmin, bmax = getClassicalSterimol(mol, origin, partner)
-    assert (L - 12.714385) < epsilon
-    assert (bmin - 3.539068) < epsilon
-    assert (bmax - 6.640342) < epsilon
+    assert np.isclose(L, 12.714385)
+    assert np.isclose(bmin, 3.539068)
+    assert np.isclose(bmax, 6.640342)
