### 2021-04-08 AstraZeneca/kallisto#31 @RMeli
Thanks for the hard work on refactoring the whole test framework! LGTM!

When excluding the CLI from the tests the coverage is still very high:
```
---------- coverage: platform darwin, python 3.7.10-final-0 ----------
Name                                 Stmts   Miss Branch BrPart  Cover   Missing
src/kallisto/__init__.py                 1      0      0      0   100%
src/kallisto/atom.py                    51      4     12      2    90%   20, 29, 32, 77, 90->exit
src/kallisto/console.py                208    208     42      0     0%   3-579
src/kallisto/data/__init__.py           10      0      2      0   100%
src/kallisto/data/alpha.py            7755      0      0      0   100%
src/kallisto/data/vdw.py                 2      0      0      0   100%
src/kallisto/grid.py                  1412      1      6      2    99%   92->exit, 106
src/kallisto/methods.py                270      5    108      6    97%   36, 58, 82, 129, 344, 403->412
src/kallisto/molecule.py               125      6     50      6    92%   56->58, 58->61, 66, 89-92, 124, 138
src/kallisto/reader/__init__.py          0      0      0      0   100%
src/kallisto/reader/strucreader.py      33     11     10      4    60%   13-22, 38, 40->45, 41->40, 53-56
src/kallisto/reader/turbomole.py        19      0      8      2    93%   23->27, 27->39
src/kallisto/reader/xyz.py              14      0      2      0   100%
src/kallisto/rmsd.py                   268     19     70      9    91%   146, 250, 302, 328->332, 364, 389->397, 412-420, 436-441, 457-458
src/kallisto/sort.py                    30     30      6      0     0%   4-66
src/kallisto/sterics.py                 41      3      4      1    87%   38-40
src/kallisto/units.py                   48      2      0      0    96%   53-54
src/kallisto/utils/__init__.py          13      6      2      0    47%   14-15, 20-21, 26-27
src/kallisto/utils/alpha.py              8      1      2      1    80%   10
TOTAL                                10308    296    324     33    96%
```
