### 2021-04-08 #3050 @RMeli
`kallisto` is a tool to calculate a series of useful atomic features. The library makes a series of published methods (detailed in the paper and the documentation, with reference) easily available in an easy to install Python package that can be used as a library or as standalone tool via a convenient CLI.

The code is well written, relatively easy to follow and well tested. I recommend publication of this work in JOSS.

---

For the future I have a few extra suggestions that I think are outside the scope of this review:
- Explore if [NetworkX](https://networkx.org/) (or similar libraries) can be used for "Breadth First Sorting" and "Substructure Finder" instead of relying on custom implementations (code you did not write is code you don't have to maintain)
- Improve docstrings by describing all input arguments so that an extensive API documentation can be built as well (the user documentation is great!)

---

Thanks @f3rmion for your great engagement with this review, despite some tasks were quite demanding. I hope our comments helped improve the package (which was already in a good state when submitted).
