### 2021-04-09 #3050 @f3rmion
Hi @RMeli, Hi @Sulstice,

thanks for your great reviews! I think that ``kallisto`` has been drastically improved since submission. It was fun to incorporate your suggestions accordingly.

> For the future I have a few extra suggestions that I think are outside the scope of this review:

> Explore if NetworkX (or similar libraries) can be used for "Breadth First Sorting" and "Substructure Finder" instead of relying on custom implementations (code you did not write is code you don't have to maintain)

> Improve docstrings by describing all input arguments so that an extensive API documentation can be built as well (the user documentation is great!)

Good points, an API documentation will follow in the near future. NetworkX looks interesting (will have a look into it).

@richardjgowers All points have been checked out :)
