### 2021-04-14 #3050 @kyleniemeyer
Hi @f3rmion, I'm the AEIC on duty this week. It looks like this submission is at the finish line - I took a look over the paper and it seems in good shape.

Can you now archive the software repository (e.g., in Zenodo) and provide the DOI here?
