### 2021-04-15 #3050 @f3rmion
> Hi @f3rmion, I'm the AEIC on duty this week. It looks like this submission is at the finish line - I took a look over the paper and it seems in good shape.
>
> Can you now archive the software repository (e.g., in Zenodo) and provide the DOI here?

Hi @kyleniemeyer,

v1.0.7 includes all changes and is now archived in Zenodo, DOI: 10.5281/zenodo.4692147
